add remaining pages
