+++ dash/app.py
@@ -1,12 +1,8 @@
 from flask import Flask
-from dash import Dash, html, dcc
+from dash import Dash, html, callback
 import dash
 import dash_bootstrap_components as dbc
-# from multipage import Route, MultiPageApp
-# from pages import Index, Series, Search, Leaderboard
-# from pages_static import Methodology, About
-
-# from blog import BlogSection
+from dash.dependencies import Input, Output, State
 from common import header, footer
 
 
@@ -27,6 +23,18 @@ app.layout = html.Div([
     dash.page_container,
     footer(), 
 ])
+
+### callback for toggling the collapse on small screens
+@callback(
+    Output("navbar-collapse", "is_open"),
+    [Input("navbar-toggler", "n_clicks")],
+    [State("navbar-collapse", "is_open")],
+)
+def toggle_navbar_collapse(n, is_open):
+    if n:
+        return not is_open
+    return is_open
+
 
 if __name__ == "__main__":
     server.run(host="0.0.0.0", port=8001, debug=True)
