+++ dash/pages/blog/post.py
@@ -33,7 +33,7 @@ from util import (
     watermark_information,
     dash_kwarg,
 )
-from modelutil import (
+from common import (
     select_best_model,
     get_thumbnail_figure,
     get_forecast_data,
