+++ dash/pages/search.py
@@ -1,42 +1,25 @@
 import json
-import pickle
 import re
 
-from datetime import datetime
-from functools import wraps
 from urllib.parse import urlencode
 
 import dash_bootstrap_components as dbc
 from dash import dcc, html, callback
 import dash
 
-import humanize
-import numpy as np
-import pandas as pd
-from dash import dash_table
 from common import breadcrumb_layout
 from dash.dependencies import Input, Output
-from dash.exceptions import PreventUpdate
-from frontmatter import Frontmatter
 from util import (
-    glob_re,
     location_ignore_null,
     parse_state,
     apply_default_value,
-    watermark_information,
     dash_kwarg,
 )
-from modelutil import (
+from common import (
     select_best_model,
     get_thumbnail_figure,
     get_forecast_data,
-    component_news_4col,
-    component_figs_2col,
 )
-
-
-import io
-import base64
 
 dash.register_page(
     __name__,
